+++ inst/shinyapp/engine/tabulate_h.R
@@ -297,7 +297,8 @@ count_col_ns <- tab |> distinct(col_index, stat_type) |> filter(stat_type == "n"
         row_index == row_index,
         dplyr::between(col_index, col_start_1, col_end_1)
       )
-    ) 
+    ) |>
+    restore_suppression_basis()
   
   out <-
     results1 |>